Wolk sensor FW wolk about platform v2.0

+++ wolksensor/SDK/core/sensors.h
@@ -43,7 +43,7 @@ sensor_t;
 typedef struct
 {
 	char id;
-	int16_t value;
+	float value;
 }
 sensor_state_t;
